part 4 update

+++ part4/blogList/bloglist-backend/controllers/blogs.js
@@ -1,12 +1,19 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const helper = require('../utils/middleware')
 
-blogsRouter.get('/', (request, response) => {
-    Blog
+blogsRouter.get('/', async (request, response) => {
+    const blogs = await Blog
         .find({})
-        .then(blogs => {
-            response.json(blogs.map(blog => blog.toJSON()))
+        .populate('user', {
+            username: 1,
+            name: 1
         })
+
+    response.json(blogs.map(b => b.toJSON()))
+
 })
 
 blogsRouter.get('/:id', async (request, response, next) => {
@@ -24,36 +31,72 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
-
-    const blog = new Blog({
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes,
-    })
+    const token = helper.tokenExtractor(request)
 
     try {
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response
+                .status(401)
+                .json({
+                    error: 'token missing or invalid'
+                })
+        }
+
+        const user = await User.findById(body.user)
+
+        const blog = new Blog({
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes,
+            user: user._id
+        })
+
         if (!blog.title || !blog.url) {
             response
                 .status(400)
                 .end()
         }
+
         if (!blog.likes) {
             blog.likes = 0
         }
+
         if (blog.title && blog.url) {
             const savedBlog = await blog.save()
+            user.blogs = user.blogs.concat(savedBlog._id)
+            await user.save()
             response.json(savedBlog.toJSON())
         }
+
     } catch (exception) {
         next(exception)
     }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const token = helper.tokenExtractor(request)
+
     try {
-        const blog = await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response
+                .status(401)
+                .json({
+                    error: 'token missing or invalid'
+                })
+        }
+
+        const blog = await Blog.findById(request.params.id)
+
+        if (blog.user.toString() === decodedToken.id) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(404).end()
+        }
+
     } catch (exception) {
         next(exception)
     }
@@ -61,18 +104,33 @@ blogsRouter.delete('/:id', async (request, response, next) => {
 
 blogsRouter.put('/:id', async (request, response, next) => {
     const body = request.body
-
-    const blog = {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes,
-    }
+    const token = helper.tokenExtractor(request)
 
     try {
-        const newBlog = await Blog
-            .findByIdAndUpdate(request.params.id, blog, { new: true })
-        response.json(newBlog.toJSON())
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) {
+            return response
+                .status(401)
+                .json({
+                    error: 'token missing or invalid'
+                })
+        }
+
+        const blog = {
+            title: body.title,
+            author: body.author,
+            url: body.url,
+            likes: body.likes,
+        }
+
+        const newBlog = await Blog.fingById(request.params.id)
+
+        if (newBlog.user.toString() === decodedToken.id) {
+            await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+            response.json(newBlog.toJSON())
+        } else {
+            response.status(404).end()
+        }
     } catch (exception) {
         next(exception)
     }
